separate string util into data conversion module, test, and isolate from staff_area test
