+++ openassessment/xblock/test/test_staff_area.py
@@ -24,7 +24,6 @@ from openassessment.workflow import api as workflow_api
 from openassessment.xblock.data_conversion import prepare_submission_for_serialization
 from openassessment.xblock.test.base import XBlockHandlerTestCase, scenario
 from openassessment.xblock.test.test_team import MockTeamsService, MOCK_TEAM_MEMBERS, MOCK_TEAM_NAME
-from openassessment.xblock.staff_area_mixin import list_to_conversational_format
 from submissions import api as sub_api
 
 FILE_URL = 'www.fileurl.com'
@@ -733,8 +732,14 @@ class TestCourseStaff(XBlockHandlerTestCase):
             'submission_returned_uuid': submission['uuid']
         })
 
+    @patch('openassessment.xblock.staff_area_mixin.list_to_conversational_format')
     @scenario('data/team_submission.xml', user_id='Bob')
-    def test_staff_form_for_team_assessment(self, xblock):
+    def test_staff_form_for_team_assessment(self, xblock, formatter):
+        # mock list format
+        def mock_formatter(entries):
+            return ', '.join(entries)
+        formatter.side_effect = mock_formatter
+
         # Simulate that we are course staff
         xblock.xmodule_runtime = self._create_mock_runtime(
             xblock.scope_ids.usage_id, True, False, "Bob"
@@ -745,7 +750,7 @@ class TestCourseStaff(XBlockHandlerTestCase):
         xblock.runtime._services['teams'] = MockTeamsService(True)  # pylint: disable=protected-access
         team_submission_enabled = 'data-team-submission="True"'
         team_name_query = 'data-team-name="{}"'.format(MOCK_TEAM_NAME)
-        team_usernames_query = 'data-team-usernames="{}"'.format(list_to_conversational_format(MOCK_TEAM_MEMBERS))
+        team_usernames_query = 'data-team-usernames="{}"'.format(mock_formatter(MOCK_TEAM_MEMBERS))
 
         # Create a submission for Bob, and corresponding workflow.
         team_item = TEAMMATE_ITEM.copy()
